Move the shared feed parser out of the playground's public directory
The monitor imported it from playground/public, which is the directory
Vite copies verbatim onto the deployed site. The parser sat there
because nothing bundles that directory, so the reader could only import
its immediate neighbours.

The reader is a second Vite entry now, which lifts that limit, and the
parser and its test live in scripts/github-changelog alongside the
other repository tooling. Both sides import from there.

The reader keeps the lint coverage it had in public: eslint ignores it.
Its 780 lines produce 310 findings under the playground's rules, which
is its own change.
