+++ package.json
@@ -19,7 +19,7 @@
     "test": "run -p test:changelog-monitor test:comment-cop test:feed test:matcher test:shellcheck-versions typecheck",
     "test:changelog-monitor": "node --test ./.github/actions/changelog-monitor/monitor.test.mjs",
     "test:comment-cop": "node --test ./.github/actions/comment-cop/comment-cop.test.mjs",
-    "test:feed": "node --test ./playground/test/feed.test.mjs",
+    "test:feed": "node --test ./scripts/github-changelog/test.mjs",
     "test:matcher": "node --test ./scripts/generate-actionlint-matcher/test.mjs",
     "test:shellcheck-versions": "node --test ./.github/actions/shellcheck-versions/shellcheck-version-lifecycle.test.mjs",
     "test:workspaces": "npm run test --workspaces",
