+++ playground/github-changelog/index.js
@@ -1,4 +1,4 @@
-import { FEED_URL, parseFeed } from './feed.mjs';
+import { FEED_URL, parseFeed } from '../../scripts/github-changelog/feed.mjs';
 
 const el = {
 	article: document.getElementById('article'),
